move root route logic to blueprint, add user repo to app
